+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### 0.4.1
+- Binary build for `linux-arm64` has been added as part of release
+- Enhanced error reporting
+
+Build Support Matrix:
+| OS \ ARCH      | AMD64       | ARM64       |
+| :------------- | :---------: | :---------: |
+| **Linux**      | ✅          | ✅          |
+| **MacOS**      |             | ✅          |
+| **Windows**    | ✅          |             |
+
 ### 0.4.0
 *Compatible with outputs from Ripple1D Pipeline version 0.10.3 to present*
 
@@ -20,10 +31,8 @@
 
 
 ### 0.2.1
-*Compatible with outputs from Ripple1D version unknown to 0.7.0*
-
 - Fix bug that was assuming no data value -9999.0 for all raster types, the no data value is now inferred from FIM library rasters
-- An error is raised in `fim` command when control file is empty
+- An error is raised in `fim` command when controls file is empty
 
 ### 0.2.0
 *Compatible with outputs from Ripple1D pipeline version unknown to 0.7.0*